# src/main/resources/openapi-facade.yml
openapi: 3.0.3
info:
  title: API Facade DDST (FAIN_J)
  description: |
    Service facade pour les instanciateurs de starter kits DDST
  version: 0.0.1-SNAPSHOT
components:
  schemas:
    Features:
      type: array
      items:
        type: string
    StarterKitType:
      type: string
      enum:
        - TONIC
        - HUMAN
        - STUMP
    ContractType:
      type: string
      enum:
        - OPENAPI
        - AVRO
        - SOAP
    ComponentTypes:
      type: array
      items:
          $ref: '#/components/schemas/StarterKitType'
paths:
  /components/starter-kits:
    get:
      tags:
        - components
      summary: Lister tous les composants disponibles
      description: Récupère la liste de tous les composants starter kit supportés
      operationId: getAllComponents
      responses:
        '200':
          description: Liste de tous les composants disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentTypes'
        '500':
          description: Erreur interne du serveur

  /components/{starter-kit}:
    get:
      tags:
        - components
      summary: Instancier un composant starter kit
      description: Créer une instance du starter kit demandé avec les fonctionnalités spécifiées
      operationId: instantiateStarterKit
      parameters:
        - name: starter-kit
          in: path
          description: Type de starter kit à instancier
          required: true
          schema:
            $ref: '#/components/schemas/StarterKitType'
        - name: features
          in: query
          description: Fonctionnalités à instancier
          required: false
          schema:
            $ref: '#/components/schemas/Features'
          style: form
          explode: true
        - name: code-applicatif
          in: query
          description: Code applicatif du composant à générer
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archive zip contenant les sources générées
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Paramètres invalides
        '500':
          description: Erreur interne du serveur

  /components/{starter-kit}/features:
    get:
      tags:
        - components
      summary: Lister les fonctionnalités d'un composant
      description: Récupère la liste des fonctionnalités disponibles pour le starter kit spécifié
      operationId: getComponentFeatures
      parameters:
        - name: starter-kit
          in: path
          description: Type de starter kit
          required: true
          schema:
            $ref: '#/components/schemas/StarterKitType'
      responses:
        '200':
          description: Liste des fonctionnalités disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Features'
        '400':
          description: Type de starter kit invalide
        '500':
          description: Erreur interne du serveur

  /contracts/{starter-kit}:
    get:
      tags:
        - contracts
      summary: Lister les contrats disponibles pour un starter kit
      description: Récupère la liste des types de contrats supportés par le starter kit spécifié
      operationId: getAvailableContracts
      parameters:
        - name: starter-kit
          in: path
          description: Type de starter kit
          required: true
          schema:
            $ref: '#/components/schemas/StarterKitType'
      responses:
        '200':
          description: Liste des contrats disponibles pour le starter kit
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContractType'
        '400':
          description: Type de starter kit invalide
        '500':
          description: Erreur interne du serveur

  /contracts/{starter-kit}/{contract-type}:
    get:
      tags:
        - contracts
      summary: Créer un contrat pour un starter kit
      description: Générer une instance de contrat pour le starter kit spécifié
      operationId: createContract
      parameters:
        - name: starter-kit
          in: path
          description: Type de starter kit
          required: true
          schema:
            $ref: '#/components/schemas/StarterKitType'
        - name: contract-type
          in: path
          description: Type de contrat à générer
          required: true
          schema:
            $ref: '#/components/schemas/ContractType'
        - name: code-applicatif
          in: query
          description: Code applicatif du composant à générer
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archive zip contenant le contrat généré
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Paramètres invalides ou combinaison starter kit/contrat non supportée
        '500':
          description: Erreur interne du serveur

  /libraries/starter-kits:
    get:
      tags:
        - libraries
      summary: Lister les starter kits disponibles pour les bibliothèques
      description: Récupère la liste des starter kits supportant la génération de bibliothèques
      operationId: getAvailableLibraryStarterKits
      responses:
        '200':
          description: Liste des starter kits disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StarterKitType'
        '500':
          description: Erreur interne du serveur

  /libraries/{starter-kit}:
    get:
      tags:
        - libraries
      summary: Créer une bibliothèque pour un starter kit
      description: Générer une instance de bibliothèque pour le starter kit spécifié
      operationId: createLibrary
      parameters:
        - name: starter-kit
          in: path
          description: Type de starter kit
          required: true
          schema:
            $ref: '#/components/schemas/StarterKitType'
        - name: code-applicatif
          in: query
          description: Code applicatif du composant à générer
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archive zip contenant la bibliothèque générée
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Paramètres invalides
        '500':
          description: Erreur interne du serveur